requirements.txt
gradio
transformers==4.46.0
pillow
requests